agrotool/PlotlyInitializing: Plotly using added in addition to maptplotlib + big bunch of updates & structure refactoring
